Do not flush on create
Previous to this commit, a flush would take place regardless if the
create method was setting parameters. Since the create method does not
set any values in the @property_flush hash, the flush fails. The flush,
however, is unnecessary.

This commit sets a @noflush instance variable in the create method that
the flush method uses to determine whether or not to run.