Do not munge array values for rules
Previous to this commit, the node_group type would munge the array
values for rules.  However, this ended up having the side effect of
combining every rule for every previously parsed node_group type.
Munging the array values in unecessary as :array_matching => :all can be
used in the property definition.

This commit removes the munging in favor of using :array_matching