Fix override_environment parameter
There is some "usability" logic for the `override_environment` parameter that
maps the parameter value to `environment_trumps` during group creation
requests. However, the logic was inverted and would perform a switch to
`override_environment` if a non-existent `environment_trumps` parameter
was in use. This caused group creation to fail as the Classifier API would
reject the `override_environment` parameter.

This patch fixes the faulty logic and cleans up some redundant statements.