Allow colons in node group names
A colon is a valid character there.  Prior to this, the node_group
resource would fail for groups with a colon in their name because the
validation regex did not allow for a colon.

This is especially relevant when node group names match a class name,
such as a 'role' class with colons for namespacing.