Do not require 'puppetclassify' to compile
There is no functional requirement for the puppetclassify gem to be
installed in order to compile a catalog that includes node_group types.
This commit modifies the code such that the compilation process will not
fail if puppetclassify is not yet installed. This enables the bootstrap
use case, where the master can in a single Puppet run install the
puppetclassify gem on itself and then use it to configure the
classifier.