Fix metaparameter bug
Previously, due to dynamism in building the JSON data to send to the NC
API, when metaparameters like "require" and "subscribe" were passed to
node_group resources those parameter keys would end up in the JSON send
to the NC - specifically in the create() method of the provider. This
resulted in the provider being unable to create resources if they had
metaparameters specified.

This commit fixes the problem and allows require, subscribe, and more to
be specified by discarding metaparameters when building the create JSON.