Fix $ngs caching when creating new groups
Previously, the $ngs global would not be populated with newly created
group IDs because of an incorrect hash index. send_data[:name] was
referrenced, which was nil, rather than send_data['name']. This made it
impossible to create a parent group and then create child groups for
that parent during the same Puppet run.

This commit fixes the problem.