Fix inheritance code for Puppet::Util::Node_groups
Previously this class did not completely inherit the functionality of
its parent, and in implemented several measures that gave it some parent
functionality but not all.

This commit modifies the inheritance code to make the child object a
fully-functional PuppetClassify object in addition to its own special
functionality, and reduces complexity and LOC at the same time.